moved all db files to mongo subfolder; server now calls imported methods, rather than invoking mongoose model's methods directly
