+++ DB/mongo/deleteData.js
